ai-service/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings

# Google Cloud & Vertex AI
google-cloud-aiplatform
google-auth
google-cloud-texttospeech
google-cloud-vision
google-cloud-storage
google-generativeai
vertexai

# RAG & Document Processing
langchain
langchain-community
chromadb
pypdf2
pypdf

# Image Processing & Document Handling
pillow

# Utilities
python-multipart
aiofiles
requests
diskcache

# Production server
gunicorn
